tests/wpt/web-platform-tests/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-border-radius-hittest.html

<!doctype html>
<title>fieldset, border-radius and hit testing</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
fieldset { width: 80px; height: 80px; border-radius: 100px; border: 10px solid; background: lime; }
</style>
<fieldset>
</fieldset>
<script>
test(() => {
  assert_equals(document.elementFromPoint(20, 20), document.body);
});
</script>
